auburn-usda-main/requirements.txt
coloredlogs==15.0.1
pandas==2.2.1
PySide6==6.6.1
PySide6==6.6.3.1
PySide6_Addons==6.6.1
PySide6_Essentials==6.6.1
PythonQwt==0.11.2
shiboken6==6.6.1
